+++ app/assets/javascripts/ui.js
@@ -2776,6 +2776,10 @@
       loop: {
         type: Boolean,
         default: true
+      },
+      autofocus: {
+        type: Boolean,
+        default: false
       }
     };
     connect() {
@@ -2789,7 +2793,7 @@
       this.element.removeEventListener("drawer:show", this.handleShow.bind(this));
     }
     handleShow() {
-      if (this.hasInputTarget) {
+      if (this.autofocusValue && this.hasInputTarget) {
         this.inputTarget.focus();
       }
       const visibleItems = this.visibleItems;