+++ app/components/ui/command.rb
@@ -3,8 +3,9 @@
 class UI::Command < Phlex::HTML
   include UI::CommandBehavior
 
-  def initialize(loop: true, classes: "", **attributes)
+  def initialize(loop: true, autofocus: false, classes: "", **attributes)
     @loop = loop
+    @autofocus = autofocus
     @classes = classes
     @attributes = attributes
   end