+++ app/components/ui/command_dialog.rb
@@ -3,8 +3,9 @@
 class UI::CommandDialog < Phlex::HTML
   include UI::CommandDialogBehavior
 
-  def initialize(shortcut: "meta+j", classes: "", **attributes)
+  def initialize(shortcut: "meta+j", autofocus: true, classes: "", **attributes)
     @shortcut = shortcut
+    @autofocus = autofocus
     @classes = classes
     @attributes = attributes
   end
@@ -14,7 +15,7 @@ class UI::CommandDialog < Phlex::HTML
       render UI::Dialog.new(close_on_escape: true, close_on_overlay_click: true) do
         render UI::DialogOverlay.new do
           render UI::DialogContent.new(classes: command_dialog_content_classes) do
-            render UI::Command.new do
+            render UI::Command.new(autofocus: @autofocus) do
               yield if block_given?
             end
           end